+++ kloppy/helpers.py
@@ -6,9 +6,11 @@ from . import (
     MetricaTrackingSerializer,
     OptaSerializer,
     SportecEventSerializer,
+    SkillCornerTrackingSerializer,
     StatsBombSerializer,
     TRACABSerializer,
     WyscoutSerializer,
+    XMLCodeSerializer,
 )
 from .domain import (
     CardEvent,
@@ -29,6 +31,8 @@ from .domain import (
     Transformer,
     Provider,
     build_coordinate_system,
+    CodeDataset,
+    Code,
 )
 
 
@@ -36,6 +40,20 @@ def load_tracab_tracking_data(
     metadata_filename: str, raw_data_filename: str, options: dict = None
 ) -> TrackingDataset:
     serializer = TRACABSerializer()
+    with open(metadata_filename, "rb") as metadata, open(
+        raw_data_filename, "rb"
+    ) as raw_data:
+
+        return serializer.deserialize(
+            inputs={"metadata": metadata, "raw_data": raw_data},
+            options=options,
+        )
+
+
+def load_skillcorner_tracking_data(
+    metadata_filename: str, raw_data_filename: str, options: dict = None
+) -> TrackingDataset:
+    serializer = SkillCornerTrackingSerializer()
     with open(metadata_filename, "rb") as metadata, open(
         raw_data_filename, "rb"
     ) as raw_data:
@@ -184,6 +202,22 @@ def load_wyscout_event_data(
         )
 
 
+def load_xml_code_data(xml_filename: str) -> CodeDataset:
+    serializer = XMLCodeSerializer()
+    with open(xml_filename, "rb") as xml_file:
+        return serializer.deserialize(inputs={"xml_file": xml_file})
+
+
+def write_xml_code_data(dataset: CodeDataset, xml_filename: str):
+    serializer = XMLCodeSerializer()
+    with open(xml_filename, "wb") as xml_file:
+        content = serializer.serialize(dataset)
+        xml_file.write(content)
+
+
+DatasetT = TypeVar("DatasetT")
+
+
 def transform(
     dataset: Dataset,
     to_orientation=None,
@@ -321,6 +355,19 @@ def _event_to_pandas_row_converter(event: Event) -> Dict:
     return row
 
 
+def _code_to_pandas_row_converter(code: Code) -> Dict:
+    row = dict(
+        code_id=code.code_id,
+        period_id=code.period.id if code.period else None,
+        timestamp=code.timestamp,
+        end_timestamp=code.end_timestamp,
+        code=code.code,
+    )
+    row.update(code.labels)
+
+    return row
+
+
 def to_pandas(
     dataset: Union[Dataset, List[DataRecord]],
     _record_converter: Callable[[DataRecord], Dict] = None,
@@ -360,7 +407,7 @@ def to_pandas(
         return pd.DataFrame()
 
     if not _record_converter:
-        if isinstance(dataset, TrackingDataset) and isinstance(
+        if isinstance(dataset, TrackingDataset) or isinstance(
             records[0], Frame
         ):
             _record_converter = _frame_to_pandas_row_converter
@@ -368,6 +415,8 @@ def to_pandas(
             records[0], Event
         ):
             _record_converter = _event_to_pandas_row_converter
+        elif isinstance(dataset, CodeDataset) or isinstance(records[0], Code):
+            _record_converter = _code_to_pandas_row_converter
         else:
             raise Exception("Don't know how to convert rows")
 
@@ -388,6 +437,7 @@ def to_pandas(
 
 __all__ = [
     "load_tracab_tracking_data",
+    "load_skillcorner_tracking_data",
     "load_metrica_tracking_data",
     "load_metrica_json_event_data",
     "load_epts_tracking_data",
@@ -395,6 +445,8 @@ __all__ = [
     "load_opta_event_data",
     "load_sportec_event_data",
     "load_wyscout_event_data",
+    "load_xml_code_data",
+    "write_xml_code_data",
     "to_pandas",
     "transform",
 ]