+++ kloppy/infra/serializers/tracking/epts/serializer.py
@@ -8,6 +8,7 @@ from kloppy.domain import (
     AttackingDirection,
     Frame,
     Point,
+    Point3D,
     Team,
     Orientation,
     Transformer,
@@ -59,7 +60,9 @@ class EPTSSerializer(TrackingDataSerializer):
             ball_state=None,
             period=period,
             players_coordinates=players_coordinates,
-            ball_coordinates=Point(x=row["ball_x"], y=row["ball_y"]),
+            ball_coordinates=Point3D(
+                x=row["ball_x"], y=row["ball_y"], z=row["ball_z"]
+            ),
         )
 
         if transformer is not None: